Completed 27/03/2015 - Javadoc attached.

+++ shopping/src/main/java/finalProject/shopping/model/Payment.java
@@ -19,7 +19,7 @@ import javax.persistence.Table;
 public abstract class Payment {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	protected int transactionNumber;
+	protected long transactionNumber;
 	
 	@Column(name="DATE")
 	protected Date date;
